fix(IR-5213): saveScene, does not ask to confirm, show toast notification instead
