+++ packages/client-core/i18n/en/editor.json
@@ -1449,7 +1449,7 @@
       "lbl-confirm": "Save Scene",
       "info-confirm": "Are you sure you want to save the scene?",
       "info-question": "Do you want to save the current scene?",
-      "info-save-success": "Scene saved successfully!",
+      "info-save-success": "Scene Saved",
       "unsavedChanges": {
         "title": "Unsaved Changes"
       }
