+++ packages/editor/src/components/toolbar/Toolbar.tsx
@@ -44,14 +44,14 @@ import { MdOutlineKeyboardArrowDown } from 'react-icons/md'
 import { RxHamburgerMenu } from 'react-icons/rx'
 import { twMerge } from 'tailwind-merge'
 import { inputFileWithAddToScene } from '../../functions/assetFunctions'
-import { onNewScene } from '../../functions/sceneFunctions'
+import { onNewScene, onSaveScene } from '../../functions/sceneFunctions'
 import { cmdOrCtrlString } from '../../functions/utils'
 import { EditorState } from '../../services/EditorServices'
 import { UIAddonsState } from '../../services/UIAddonsState'
 import CreatePrefabPanel from '../dialogs/CreatePrefabPanelDialog'
 import CreateSceneDialog from '../dialogs/CreateScenePanelDialog'
 import ImportSettingsPanel from '../dialogs/ImportSettingsPanelDialog'
-import { SaveNewSceneDialog, SaveSceneDialog } from '../dialogs/SaveSceneDialog'
+import SaveNewSceneDialog from '../dialogs/SaveNewSceneDialog'
 
 const onImportAsset = async () => {
   const { projectName } = getState(EditorState)
@@ -69,11 +69,7 @@ export const confirmSceneSaveIfModified = async () => {
   const isModified = EditorState.isModified()
 
   if (isModified) {
-    return new Promise((resolve) => {
-      PopoverState.showPopupover(
-        <SaveSceneDialog isExiting onConfirm={() => resolve(true)} onCancel={() => resolve(false)} />
-      )
-    })
+    return onSaveScene()
   }
   return true
 }
@@ -121,7 +117,7 @@ const generateToolbarMenu = () => {
     {
       name: t('editor:menubar.saveScene'),
       hotkey: `${cmdOrCtrlString}+s`,
-      action: () => PopoverState.showPopupover(<SaveSceneDialog />)
+      action: onSaveScene
     },
     {
       name: t('editor:menubar.saveAs'),
